refactor: split reading (#42)
